move plugin to src folder and fix lint issues
